+++ tools/src/main/java/org/apache/kafka/tools/ConsumerPerformance.java
@@ -31,7 +31,6 @@ import org.apache.kafka.common.utils.Exit;
 import org.apache.kafka.common.utils.Utils;
 import org.apache.kafka.server.util.CommandDefaultOptions;
 import org.apache.kafka.server.util.CommandLineUtils;
-import org.apache.kafka.server.util.ToolsUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
