+++ tools/src/main/java/org/apache/kafka/tools/ProducerPerformance.java
@@ -44,7 +44,6 @@ import net.sourceforge.argparse4j.inf.Namespace;
 import org.apache.kafka.common.utils.Exit;
 import org.apache.kafka.common.utils.ThroughputThrottler;
 import org.apache.kafka.common.utils.Utils;
-import org.apache.kafka.server.util.ToolsUtils;
 
 public class ProducerPerformance {
